requirements.txt
# PubMed/PMC drug context agent — Python 3.10+
mcp>=1.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
requests>=2.31.0
rich>=13.0.0
httpx>=0.27.0
pandas>=2.0.0
openpyxl>=3.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
